aggiunto risorsa create

+++ src/app/features/risorsa/risorsa-delete/risorsa-delete.component.ts
@@ -26,7 +26,7 @@ export class RisorsaDeleteComponent {
   }
 
   delete() {
-    if (this.risorsaDelete) {
+    if (this.risorsaDelete&&this.risorsaDelete.id) {
       this.risorsaService.removeRisorsa(this.risorsaDelete.id).subscribe(
         () => {
           this.router.navigate(['/risorsa/list']);
